requirements.txt
django-tagging
hadrian
django-ckeditor